Add relative API
