+++ relative.js
@@ -1,0 +1,4 @@
+/* eslint-env node */
+
+// This is a fallback for older node that doesn't know about package.json's exports field.
+module.exports = require('./dist/relative.cjs.js');
